exception handler deleteById()

+++ java-springboot/springBootCommerce/src/main/java/com/backendPortfolio/springBootCommerce/services/UserService.java
@@ -1,12 +1,16 @@
 package com.backendPortfolio.springBootCommerce.services;
 
+
 import com.backendPortfolio.springBootCommerce.entities.User;
 import com.backendPortfolio.springBootCommerce.repositories.UserRepository;
-import com.backendPortfolio.springBootCommerce.services.exceptions.ResourceNotFindException;
+import com.backendPortfolio.springBootCommerce.services.exceptions.DatabaseException;
+import com.backendPortfolio.springBootCommerce.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.security.SecurityProperties;
-import org.springframework.stereotype.Component;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+
+import jakarta.persistence.EntityNotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,29 +20,47 @@ public class UserService {
 
     @Autowired
     private UserRepository repository;
-    public List<User> findAll(){
+
+    public List<User> findAll() {
         return repository.findAll();
     }
-    public User findById(Long id){
-        Optional<User> obj =  repository.findById(id);
-        return obj.orElseThrow(()->new ResourceNotFindException(id));
+
+    public User findById(Long id) {
+        Optional<User> obj = repository.findById(id);
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
-    public User insert(User obj){
+
+    public User insert(User obj) {
         return repository.save(obj);
     }
-    public void delete(Long id){
-        repository.deleteById(id);
+
+    public void delete(Long id) {
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        } else {
+            try {
+                repository.deleteById(id);
+            } catch (DataIntegrityViolationException e) {
+                throw new DatabaseException(e.getMessage());
+            }
+        }
     }
-    public  User update(Long id, User obj){
-        User entity = repository.getReferenceById(id);
-        updateData(entity, obj);
-        return repository.save(entity);
+
+
+    public User update(Long id, User obj) {
+        try {
+            User entity = repository.getReferenceById(id);
+            updateData(entity, obj);
+            return repository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(User entity, User obj) {
-        entity.setNome(obj.getNome());
+        entity.setName(obj.getName());
         entity.setEmail(obj.getEmail());
         entity.setPhone(obj.getPhone());
     }
-
 }
